+++ saleor/graphql/app/tests/benchmarks/test_apps.py
@@ -23,7 +23,7 @@ def test_apps_for_federation_query_count(
     )
 
     query = """
-        query GetAppInFederation($representations: [_Any]) {
+        query GetAppInFederation($representations: [_Any!]!) {
             _entities(representations: $representations) {
                 __typename
                 ... on App {